enscrap/pyproject.toml
[project]
name = "enscrap"
version = "0.1.0"
description = "Enscrap using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "pydantic>=2.11.0,<3.0.0",
    "supabase>=2.7.1,<3.0.0",
    "requests>=2.32.3,<3.0.0",
    "beautifulsoup4>=4.12.3,<5.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pyyaml>=6.0.2,<7.0.0"
]

[project.scripts]
enscrap = "enscrap.main:run"
run_crew = "enscrap.main:run"
train = "enscrap.main:train"
replay = "enscrap.main:replay"
test = "enscrap.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.hatch.build.targets.wheel]
packages = ["src/enscrap"]